Adds math module for salesanalyst
