+++ lib/sales_analyst.rb
@@ -1,7 +1,10 @@
 require 'bigdecimal'
 require 'bigdecimal/util'
+require_relative '../lib/math'
 
 class SalesAnalyst
+  include Math
+  
   def initialize(sales_engine)
     @item_repo = sales_engine.items
     @merchant_repo = sales_engine.merchants
@@ -238,6 +241,8 @@ class SalesAnalyst
     pending_merchant_ids.uniq.map do |id|
       @merchant_repo.find_by_id(id)
     end
+  end
+  
   def merchants_with_only_one_item
     items_per_merchant = @merchant_repo.all.map do |merchant|
       @item_repo.find_all_by_merchant_id(merchant.id)
@@ -264,28 +269,5 @@ class SalesAnalyst
       invoice_item.unit_price * invoice_item.quantity
     end
     sum merchants_invoice_items
-  end
-
-  # maths
-
-  def sum(nums)
-    nums.inject(0) do |running_count, item|
-      running_count + item
-    end
-  end
-
-  def mean(nums)
-    sum = sum(nums)
-    (sum.to_f / nums.length).round(2).to_d
-  end
-
-  def std_dev(nums)
-    mean = mean(nums)
-    nums = nums.map do |num|
-      (num - mean) * (num - mean)
-    end
-    nums_sum = sum(nums)
-    variance = nums_sum.to_f / (nums.length - 1)
-    Math.sqrt(variance).round(2)
   end
 end
